requirements.txt
Django==3.0.8
django-filter==2.3.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
psycopg2-binary==2.8.5 
PyJWT==1.7.1
pytest==5.4.1
pytest-django==3.9.0
pytz==2020.1
